More dynamic versions
This commit implements more dynamic terraform versions by downloading
the terraform sumfile and verifying the downloaded binaries through this
sumfile instead. The major issue still is that the sumfile also needs a
sum. Technically, terraform releases a signature that's signed with
their GPG key, but verifying that it's tricky.

For now, either the sum is pre-registered or supplied by the caller of
`terraform_dependencies`